wszystko

+++ com/maleckyyautokomis/Mechanicy.java
@@ -3,7 +3,7 @@ package com.maleckyyautokomis;
 public class Mechanicy {
 
     String imie;
-    int cenaMechanika;
+    double cenaMechanika;
     int skutecznosc;
